feat: add basic offline mode

+++ src/redux/features/recipesSlice.ts
@@ -2,6 +2,7 @@ import {createAsyncThunk, createSlice} from '@reduxjs/toolkit';
 import {AxiosError} from 'axios';
 import RestAPI, {Recipe, RecipeGroup} from '../../dao/RestAPI';
 import {RootState} from '../store';
+import AppPersistence from '../../AppPersistence';
 
 
 export interface RecipesState {
@@ -17,9 +18,17 @@ const initialState: RecipesState = {
   pendingRequests: 0,
 };
 
-export const fetchMyRecipes = createAsyncThunk(
+export const fetchMyRecipes = createAsyncThunk<Recipe[], void, {state: RootState}>(
     'fetchMyRecipes',
-    async (): Promise<Recipe[]> => {
+    async (_, {getState}): Promise<Recipe[]> => {
+      if (!getState().settings.isOnline) {
+        const offlineRecipe = (await AppPersistence.getRecipesOffline());
+        if (offlineRecipe === undefined) {
+          throw Error('No offline data');
+        }
+        return offlineRecipe;
+      }
+
       return RestAPI.getRecipes();
     },
 );
@@ -31,37 +40,64 @@ export const fetchSingleRecipe = createAsyncThunk<Recipe, number, { state: RootS
         return getState().recipes.recipes.filter((recipe) => recipe.id === recipeId)[0];
       }
 
+      if (!getState().settings.isOnline) {
+        const offlineRecipe = (await AppPersistence.getRecipesOffline()).filter((recipe) => recipe.id === recipeId)[0];
+        if (offlineRecipe === undefined) {
+          throw Error('No offline data');
+        }
+        return offlineRecipe;
+      }
+
       return RestAPI.getRecipeById(recipeId);
     },
 );
-export const fetchMyRecipeGroups = createAsyncThunk(
+export const fetchMyRecipeGroups = createAsyncThunk<RecipeGroup[], any, {state: RootState}>(
     'fetchMyRecipeGroups',
-    async (): Promise<RecipeGroup[]> => {
+    async (_, {getState}): Promise<RecipeGroup[]> => {
+      if (!getState().settings.isOnline) {
+        const offlineRecipeGroups = (await AppPersistence.getRecipeGroupsOffline());
+        if (offlineRecipeGroups === undefined) {
+          throw Error('No offline data');
+        }
+        return offlineRecipeGroups;
+      }
       return RestAPI.getRecipeGroups();
     },
 );
 
 export const createRecipeGroup = createAsyncThunk<RecipeGroup, RecipeGroup, { state: RootState }>(
     'createRecipeGroup',
-    async (recipeGroup: RecipeGroup): Promise<RecipeGroup> => {
+    async (recipeGroup: RecipeGroup, {getState}): Promise<RecipeGroup> => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       return RestAPI.createNewRecipeGroup(recipeGroup);
     },
 );
 export const updateRecipeGroup = createAsyncThunk<RecipeGroup, RecipeGroup, { state: RootState }>(
     'updateRecipeGroup',
-    async (recipeGroup: RecipeGroup): Promise<RecipeGroup> => {
+    async (recipeGroup: RecipeGroup, {getState}): Promise<RecipeGroup> => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       return RestAPI.updateRecipeGroup(recipeGroup);
     },
 );
 export const deleteRecipeGroup = createAsyncThunk<void, number, { state: RootState }>(
     'deleteRecipeGroup',
-    async (groupId: number) => {
+    async (groupId: number, {getState}) => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       RestAPI.deleteRecipeGroup(groupId);
     },
 );
 export const updateRecipe = createAsyncThunk<Recipe, Recipe, { state: RootState }>(
     'updateRecipe',
-    async (recipe: Recipe): Promise<Recipe> => {
+    async (recipe: Recipe, {getState}): Promise<Recipe> => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       return RestAPI.updateRecipe(recipe);
     },
 );
@@ -69,6 +105,9 @@ export const updateRecipe = createAsyncThunk<Recipe, Recipe, { state: RootState 
 export const importRecipe = createAsyncThunk<Recipe, string, { state: RootState, rejectValue: AxiosError }>(
     'importRecipe',
     async (importURL: string, thunk) => {
+      if (!thunk.getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       try {
         return await RestAPI.importRecipe(importURL);
       } catch (e) {
@@ -79,13 +118,19 @@ export const importRecipe = createAsyncThunk<Recipe, string, { state: RootState,
 
 export const createRecipe = createAsyncThunk<Recipe, Recipe, { state: RootState }>(
     'createRecipe',
-    async (recipe: Recipe): Promise<Recipe> => {
+    async (recipe: Recipe, {getState}): Promise<Recipe> => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       return RestAPI.createNewRecipe(recipe);
     },
 );
 export const deleteRecipe = createAsyncThunk<void, Recipe, { state: RootState }>(
     'deleteRecipe',
-    async (recipe: Recipe): Promise<void> => {
+    async (recipe: Recipe, {getState}): Promise<void> => {
+      if (!getState().settings.isOnline) {
+        throw Error('Action not possible while offline');
+      }
       return RestAPI.deleteRecipe(recipe);
     },
 );
@@ -107,6 +152,7 @@ export const recipesSlice = createSlice({
         .addCase(fetchMyRecipes.fulfilled, (state, action) => {
           state.pendingRequests--;
           state.recipes = action.payload;
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(fetchMyRecipes.rejected, (state) => {
           state.pendingRequests--;
@@ -119,6 +165,7 @@ export const recipesSlice = createSlice({
         .addCase(fetchMyRecipeGroups.fulfilled, (state, action) => {
           state.pendingRequests--;
           state.recipeGroups = action.payload;
+          AppPersistence.storeRecipeGroupsOffline(action.payload);
         })
         .addCase(fetchMyRecipeGroups.rejected, (state) => {
           state.pendingRequests--;
@@ -140,6 +187,7 @@ export const recipesSlice = createSlice({
               state.recipes[index] = action.payload;
             }
           });
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(fetchSingleRecipe.rejected, (state) => {
           state.pendingRequests--;
@@ -152,6 +200,7 @@ export const recipesSlice = createSlice({
         .addCase(createRecipe.fulfilled, (state, action) => {
           state.pendingRequests--;
           state.recipes.push(action.payload);
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(createRecipe.rejected, (state) => {
           state.pendingRequests--;
@@ -163,6 +212,7 @@ export const recipesSlice = createSlice({
         .addCase(createRecipeGroup.fulfilled, (state, action) => {
           state.pendingRequests--;
           state.recipeGroups.push(action.payload);
+          AppPersistence.storeRecipeGroupsOffline(state.recipeGroups);
         })
         .addCase(createRecipeGroup.rejected, (state) => {
           state.pendingRequests--;
@@ -179,6 +229,7 @@ export const recipesSlice = createSlice({
               state.recipes.splice(index, 1);
             }
           });
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(deleteRecipe.rejected, (state) => {
           state.pendingRequests--;
@@ -195,6 +246,7 @@ export const recipesSlice = createSlice({
               state.recipeGroups.splice(index, 1);
             }
           });
+          AppPersistence.storeRecipeGroupsOffline(state.recipeGroups);
         })
         .addCase(deleteRecipeGroup.rejected, (state) => {
           state.pendingRequests--;
@@ -211,6 +263,7 @@ export const recipesSlice = createSlice({
               state.recipes[index] = action.payload;
             }
           });
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(updateRecipe.rejected, (state) => {
           state.pendingRequests--;
@@ -226,6 +279,7 @@ export const recipesSlice = createSlice({
               state.recipeGroups[index] = action.payload;
             }
           });
+          AppPersistence.storeRecipeGroupsOffline(state.recipeGroups);
         })
         .addCase(updateRecipeGroup.rejected, (state) => {
           state.pendingRequests--;
@@ -237,6 +291,8 @@ export const recipesSlice = createSlice({
         .addCase(importRecipe.fulfilled, (state, action) => {
           state.pendingRequests--;
           state.recipeGroups.push(action.payload);
+          AppPersistence.storeRecipeGroupsOffline(state.recipeGroups);
+          AppPersistence.storeRecipesOffline(state.recipes);
         })
         .addCase(importRecipe.rejected, (state, action) => {
           state.pendingRequests--;
